requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
accelerate=1.2.1=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.10=pypi_0
aiosignal=1.3.2=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anyio=4.7.0=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=24.3.0=pypi_0
audioread=3.0.1=pypi_0
bitarray=3.0.0=pypi_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
blis=1.1.0=pypi_0
boto3=1.36.1=pypi_0
botocore=1.36.1=pypi_0
braceexpand=0.1.7=pypi_0
brotli-python=1.0.9=py39h6a678d5_9
ca-certificates=2024.12.31=h06a4308_0
cachetools=5.5.0=pypi_0
catalogue=2.0.10=pypi_0
certifi=2024.12.14=py39h06a4308_0
cffi=1.17.1=pypi_0
cfgv=3.4.0=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cloudpathlib=0.20.0=pypi_0
cloudpickle=3.1.0=pypi_0
colorama=0.4.6=pypi_0
confection=0.1.5=pypi_0
contexttimer=0.3.3=pypi_0
contourpy=1.3.0=pypi_0
crcmod=1.7=pypi_0
cryptography=44.0.0=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.10=pypi_0
cython=3.0.11=pypi_0
datasets=3.0.1=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
diffusers=0.31.0=pypi_0
dill=0.3.8=pypi_0
distlib=0.3.9=pypi_0
distro=1.9.0=pypi_0
einops=0.8.0=pypi_0
evaluate=0.4.3=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
fairscale=0.4.4=pypi_0
fairseq=0.12.2=pypi_0
filelock=3.16.1=pypi_0
filetype=1.2.0=pypi_0
fire=0.7.0=pypi_0
flatbuffers=24.12.23=pypi_0
fonttools=4.55.3=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.3.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
groundingdino-py=0.4.0=pypi_0
grpcio=1.68.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.24.7=pypi_0
hydra-core=1.0.7=pypi_0
identify=2.6.3=pypi_0
idna=3.7=py39h06a4308_0
imageio=2.36.1=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
iopath=0.1.10=pypi_0
ipython=8.18.1=pypi_0
jax=0.4.30=pypi_0
jaxlib=0.4.30=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.4=pypi_0
jiter=0.8.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
kaggle=1.6.17=pypi_0
kiwisolver=1.4.7=pypi_0
langcodes=3.5.0=pypi_0
language-data=1.3.0=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
librosa=0.10.1=pypi_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.11.9=pypi_0
llvmlite=0.43.0=pypi_0
lxml=5.3.0=pypi_0
marisa-trie=1.2.1=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.10.20=pypi_0
ml-dtypes=0.5.1=pypi_0
modelscope=1.21.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
murmurhash=1.0.11=pypi_0
narwhals=1.18.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
nltk=3.9.1=pypi_0
nodeenv=1.9.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
omegaconf=2.0.6=pypi_0
openai=1.58.1=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
opendatasets=0.1.22=pypi_0
openssl=3.0.15=h5eee18b_0
opt-einsum=3.4.0=pypi_0
oss2=2.19.1=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=11.0.0=pypi_0
pip=24.0=pypi_0
platformdirs=4.3.6=pypi_0
plotly=5.24.1=pypi_0
pooch=1.8.2=pypi_0
portalocker=3.0.0=pypi_0
pre-commit=4.0.1=pypi_0
preshed=3.0.9=pypi_0
prompt-toolkit=3.0.48=pypi_0
promptcap=1.0.3=pypi_0
propcache=0.2.1=pypi_0
protobuf=4.25.5=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=18.1.0=pypi_0
pybboxes=0.1.6=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pydeck=0.9.1=pypi_0
pydeprecate=0.3.2=pypi_0
pydot=3.0.3=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.21=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-magic=0.4.27=pypi_0
python-slugify=8.0.4=pypi_0
python_abi=3.9=2_cp39
pytorch-lightning=1.7.7=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
rapidfuzz=3.11.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=py39h5eee18b_0
requests=2.32.3=py39h06a4308_1
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
roboflow=1.1.50=pypi_0
rouge-score=0.0.4=pypi_0
rpds-py=0.22.3=pypi_0
s3transfer=0.11.1=pypi_0
sacrebleu=2.4.3=pypi_0
safetensors=0.4.5=pypi_0
sahi=0.11.20=pypi_0
salesforce-lavis=1.0.0=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=69.5.1=pypi_0
shapely=2.0.6=pypi_0
shellingham=1.5.4=pypi_0
simplejson=3.19.3=pypi_0
six=1.17.0=pypi_0
smart-open=7.1.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
sounddevice=0.5.1=pypi_0
soundfile=0.12.1=pypi_0
soxr=0.5.0.post1=pypi_0
spacy=3.8.3=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.45.3=h5eee18b_0
srsly=2.5.0=pypi_0
stack-data=0.6.3=pypi_0
streamlit=1.41.1=pypi_0
supervision=0.6.0=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
taming-transformers-rom1504=0.0.6=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
terminaltables=3.1.10=pypi_0
text-unidecode=1.3=pypi_0
thinc=8.3.3=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.5.0=pypi_0
tifascore=1.1.0=dev_0
tifffile=2024.8.30=pypi_0
tiktoken=0.8.0=py39h37766c7_1
timm=0.4.12=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.1=pypi_0
torchaudio=2.5.1=pypi_0
torchmetrics=1.6.0=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.47.1=pypi_0
triton=3.1.0=pypi_0
typer=0.15.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
ultralytics=8.3.62=pypi_0
ultralytics-thop=2.0.14=pypi_0
unicodedata2=15.1.0=pypi_0
urllib3=1.26.20=pypi_0
virtualenv=20.28.0=pypi_0
wasabi=1.1.3=pypi_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.4.1=pypi_0
webdataset=0.2.100=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py39h06a4308_0
word2number=1.1=pypi_0
wrapt=1.17.0=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h5eee18b_1
yapf=0.43.0=pypi_0
yarl=1.18.3=pypi_0
yolov5=7.0.14=pypi_0
zhconv=1.4.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
